move tag and comment's action to new class, and finished comments show!

+++ trunk/src/com/twitpic/services/impl/ImageServiceGIF.java
@@ -30,7 +30,7 @@ public class ImageServiceGIF implements ImageService {
 			} else {// we make image's width to be given, the height will be
 					// changed relative
 				if (_width > width) {
-					height = (int) Math.ceil(_height * width / _width);
+					height = (int) Math.ceil((float)_height * width / _width);
 				} else {
 					return false;
 				}
